C3.Line.update can reset the `kind` of a segment
It's possible to replace (`update`) a segment using fresh data and provide
a new setting for `kind`. Previously the code didn't send the `kind` to
C3, so it would use the default (straight lines).

Modify the example to demonstrate this behaviour (using `update`
which now works rather than `flow` which is still broken).

Note the 2 definitions `base` and `range` in the example are @yminsky's
example code in #9.

Fixes #9

Signed-off-by: David Scott <dave@recoil.org>